DoodleDict/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.10.1
django-xff==1.3.0
Pillow==8.3.2
googletrans==3.1.0a0
requests==2.26.0
pygeoip==0.3.2
gunicorn==20.0.4
numpy==1.21.4
